Move .ui and compiled ui files to layout directory
