+++ src/molara/gui/layouts/measuring_tool.ui
