+++ src/molara/gui/layouts/mos_dialog.ui
