+++ src/molara/gui/layouts/normalization_dialog.ui
