+++ src/molara/gui/layouts/structure_customizer.ui
