+++ src/molara/gui/layouts/trajectory.ui
