+++ src/molara/gui/layouts/ui_normalization_dialog.py
