+++ src/molara/gui/main_window.py
@@ -19,7 +19,7 @@ from molara.gui.structure_customizer_dialog import StructureCustomizerDialog
 from molara.gui.supercell_dialog import SupercellDialog
 from molara.gui.surface_3d_dialog import CubeFileDialog
 from molara.gui.trajectory_dialog import TrajectoryDialog
-from molara.gui.ui_form import Ui_MainWindow
+from molara.gui.layouts.ui_form import Ui_MainWindow
 from molara.structure.crystal import Crystal
 from molara.structure.crystals import Crystals
 from molara.structure.io.exporter import GeneralExporter
