+++ src/molara/gui/measuring_tool_dialog.py
@@ -9,7 +9,7 @@ from PySide6.QtCore import Qt
 from PySide6.QtGui import QBrush, QColor
 from PySide6.QtWidgets import QDialog, QHeaderView, QMainWindow, QTableWidget, QTableWidgetItem
 
-from molara.gui.ui_measuring_tool import Ui_measuring_tool
+from molara.gui.layouts.ui_measuring_tool import Ui_measuring_tool
 
 if TYPE_CHECKING:
     from molara.structure.structure import Structure
